Remove the _sender helper and restore previous account-from functionality. Much better performance.
