--- src/content/helpers/_sender-parent.js
@@ -1,6 +1,0 @@
-import store from '../../store/store-content.js'
-import {respond} from '../sandbox/sandbox-messenger-server.js'
-
-respond('helper-_sender-account', () => {
-  return store.getAccount()
-})
